went long

+++ dsd/dsd.h
@@ -155,9 +155,9 @@ typedef struct
   int carrier;
   char tg[25][16];
   int tgcount;
-  int lasttg;
-  int lastsrc;
-  int src_list[50];
+  long lasttg;
+  long lastsrc;
+  long src_list[50];
   int nac;
   int errs;
   int errs2;
