+++ dsd/dsd_main.c
@@ -150,7 +150,7 @@ initState (dsd_state * state)
   state->audio_out_buf_p = state->audio_out_buf + 100;
   state->audio_out_float_buf = malloc (sizeof (float) * 1000000);
   memset (state->audio_out_float_buf, 0, 100 * sizeof (float));
-  memset (state->src_list, 0, sizeof (int) * 50);
+  memset (state->src_list, 0, sizeof (long) * 50);
   state->audio_out_float_buf_p = state->audio_out_float_buf + 100;
   state->audio_out_idx = 0;
   state->audio_out_idx2 = 0;
